backend/requirements.txt
selenium>=4.15.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
fastapi>=0.104.0
uvicorn>=0.24.0
slowapi
python-dotenv
psycopg2-binary
httpx
requests
gunicorn
uvicorn>=0.24.0
